requirements.txt
einops
h5py
matplotlib
numpy
omegaconf
lightning
Pillow
tokenizers
torch
tqdm
wandb
torchvision
scikit-learn
scipy==1.12.0
pandas
timm
PyMovieDb
transformers
python-dotenv
torchmetrics
setuptools
tensorboard
pyyaml-include
pytreebank

pytest
softadapt
PyYAML
gensim
ipython
nltk
torchaudio